Removes all hackney calls
Replaces all hackney calls in dataloader module with
httpc based client.

Tested a Man in The Middle attach and the ssl cert thing
seems to work.

There was a concious effort to change as little code as possible
in data loader but I made the call to switch the header filter
to charlists rather than iterate over the httpc header return
values as that seemed silly.