Limit applied rules to zone line boundary and...
... fix a number of edge cases.

Previously, when generating periods, when each zone line was processed,
the corresponding rules would be used to generate time periods according
to the range of years they encompassed.  This meant that any two zone
lines that switched between, for instance, daylight savings rules, could
potentially have periods generated that overlap the time period of prior
and subsequent time zones.  This caused some anomalies, such as periods
that went back in time, and nil periods, which would be cleaned up in a
post-process.

The algorithm has been modified to generate saner output which requires
no post-process step. During the generation of periods for each zone
line, the lower-limit is inferred from the end of the prior zone line,
and the upper limit is calculated for each new time period inferred from
the associated rule set of the zone line.  The rule set is followed
until a time period exhausts the upper limit.  If a period intersects
with either the lower or upper limit, the period's own lower and upper
limit will be truncated to fit inside those limits.

This means that the output should now generate periods where the
following conditions are met:

- the 'from' of a period will equal the 'until' of the prior period
- the 'until' of a period will equal the 'from' of the next period
- the 'until' of a period will always be greater than the 'from'

Thus there will be no overlaps.  In the case that a period is generated
with a matching 'from' and 'until', perhaps due to truncation, that
period will not be added to the list.

A test has been made to ensure these invariants hold for a number of
special cases found through the tzdata from `2020a`.  One of which has
disappeared from later versions of the tzdata, and so it makes sense to
keep this data separate.  An extra test function has been added to
permit parsing a single file outside of the usual test fixture.

During testing of the new algorithm, a bug was also found with the
ordering of rules.  Rule changes were ordered at month granularity, but
this causes issues where rule changes happen several times in a month,
e.g., during religious festivals.  The code has been updated to
calculate the day onset for each rule (which may change according to the
year) before sorting.