feat: support running on a read-only filesystem
Docker containers can run on a read-only filesystem, which has some
security benefits. The current updater implementation writes the .tgz
file to the filesystem, so it does not work in such an environment.

This provides a `read_only_fs?` configuration which stores the data in
memory while loading, and does not write the ETS table to disk for
future loading.